chore: CardResponse에서 CardCollectionDto로 이름 변경 및 CardCollectionDto로 makeCollection 메소드를 옮김- bn-tw2020#46
